Refractor frontend for selecting difficulty
